model_trainer/requirements.txt
pandas
scikit-learn
psycopg2-binary
boto3
sqlalchemy
matplotlib
seaborn
scikit-learn
pycaret
python-dotenv